refactor(Hand): 결과를 구하는 합계 로직(`getIncludingAceSum`)을 depth 1이 되게끔 변경
